// MH-api/src/modules/staffs/staffs.interface.ts
export interface IStaffs {
  id?: string;
  unitId?: string;
  status?: string;
  userId?: string;
  staffCode: string;
  fullName: string;
  departmentId: string;
  position?: string;
  gender: string;
  dayOfBirth: Date;
  placeOfBirth: string;
  temporaryAddress: string;
  permanentAddress: string;
  ethnic: string;
  religion: string;
  nationality: string;
  level: string;
  marital: string;
  element?: string;
  email: string;
  phoneNumber: string;
  phoneCode: string;
  peopleId: string;
  issueDate: Date;
  issuePlace: string;
  region: string;
  taxCode?: string;
  bankAccountNumber?: string;
  bankCode?: string;
  socialInsuranceId?: string;
  healthInsuranceId?: string;
  unionBookNumber?: string;
  insuranceParticipationDate?: Date;
  issueInsuranceDate?: Date;
}